Writingtofiles
Shows how to add names to a file. Uses append ('a') so that when the code is run, again and again, each new name is entered into the file and doesn't override the old names. If a 'w' was used instead of an 'a', the old names would be overwritten.