mathsgame
A maths game. Tries to read highscore from a file. If no file exists, it makes one (using try/except). Asks player questions and logs their points in score variable. Checks if score is greater than highscore. If so, it sets score as highscore. Highscore is then written to the text file to be stored for next time.